pyproject.toml
[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "data_validator"
version = "0.1.0"
description = "A flexible data validation module using YAML configuration for Spark DataFrames with support for multiple compute engines"
authors = ["Data Validator Team"]
readme = "README.md"
packages = [{include = "data_validator", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8"
pyyaml = ">=6.0"
pydantic = ">=2.0"
pyspark = {version = ">=3.3.0", optional = true}
duckdb = {version = ">=0.8.0", optional = true}
pandas = {version = ">=1.5.0", optional = true}
polars = {version = ">=0.18.0", optional = true}
"databricks-labs-dqx" = {version = ">=0.1.0", python = ">=3.10", optional = true}

[tool.poetry.extras]
spark = ["pyspark"]
duckdb = ["duckdb", "pandas"]
polars = ["polars"]
all = ["pyspark", "duckdb", "pandas", "polars"]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"
black = ">=23.0"
ruff = ">=0.1.0"
mypy = ">=1.0"

[tool.poetry.scripts]
data-validator-job = "data_validator.cli:main"

[tool.poetry.urls]
homepage = "https://github.com/infinit3labs/data_validator"
repository = "https://github.com/infinit3labs/data_validator"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "N", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]